+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Fix corrupt backup.zip because output file is not closed in backup CLI command
 * Add log message when backup is finished
+* Add command `createapitoken`
 
 
 ## v2025.12 - 2025-01-27